submits/97222061/project3/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.11.4
joblib==1.1.0
Khayyam==3.0.17
Markdown==3.3.7
numpy==1.23.0
pandas==1.4.3
prometheus-client==0.14.1
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tabulate==0.8.10
threadpoolctl==3.1.0
zipp==3.8.0
